+++ osmnx/stats.py
@@ -12,8 +12,10 @@ periphery of the graph.
 You can use NetworkX directly for additional topological network measures.
 """
 
+import itertools
 import logging as lg
 import warnings
+from collections import Counter
 
 import networkx as nx
 import numpy as np
@@ -249,6 +251,57 @@ def circuity_avg(Gu):
         return edge_length_total(Gu) / sl_dists_total
     except ZeroDivisionError:
         return None
+
+
+def count_streets_per_node(G, nodes=None):
+    """
+    Count how many physical street segments connect to each node in a graph.
+
+    This function uses an undirected representation of the graph and special
+    handling of self-loops to accurately count physical streets rather than
+    directed edges. Note: this function is automatically run by all the
+    `graph.graph_from_x` functions prior to truncating the graph to the
+    requested boundaries, to add accurate `street_count` attributes to each
+    node even if some of its neighbors are outside the requested graph
+    boundaries.
+
+    Parameters
+    ----------
+    G : networkx.MultiDiGraph
+        input graph
+    nodes : list
+        which node IDs to get counts for. if None, use all graph nodes,
+        otherwise calculate counts only for these node IDs
+
+    Returns
+    -------
+    streets_per_node : dict
+        counts of how many physical streets connect to each node, with keys =
+        node ids and values = counts
+    """
+    if nodes is None:
+        nodes = G.nodes
+
+    # get one copy of each self-loop edge, because bi-directional self-loops
+    # appear twice in the undirected graph (u,v,0 and u,v,1 where u=v), but
+    # one-way self-loops will appear only once
+    Gu = G.to_undirected(reciprocal=False, as_view=True)
+    self_loop_edges = set(nx.selfloop_edges(Gu))
+
+    # get all non-self-loop undirected edges, including parallel edges
+    non_self_loop_edges = [e for e in Gu.edges(keys=False) if e not in self_loop_edges]
+
+    # make list of all unique edges including each parallel edge unless the
+    # parallel edge is a self-loop, in which case we don't double-count it
+    all_unique_edges = non_self_loop_edges + list(self_loop_edges)
+
+    # flatten list of (u, v) edge tuples to count how often each node appears
+    edges_flat = itertools.chain.from_iterable(all_unique_edges)
+    counts = Counter(edges_flat)
+    streets_per_node = {node: counts[node] for node in nodes}
+
+    utils.log("Counted undirected street segments incident on each node")
+    return streets_per_node
 
 
 def basic_stats(